Step 6: Enabled XFT support on Cinelerra
Xft render antialiased fonts for the interface. The problem with Cinelerra was that
after some drawing, XFT crashed on 32 bit, and Cinelerra went back to un-antialiased
fonts. I think that can be solved. I believe it is a problem with freeing xft allocated
memory. For now this commit works without any problems on 64 bit and can be disabled
by --enable-xft=no on configure.

There really is no point in using Utf8 without xft. Xft provides a quick interface to
fontconfig that autoselects the correct locale font on the fly; otherwise, Cinelerra
needs to import/adapt some function, similar to this code, for bcwindows.

Antialias xft font rendering can be disabled on bcresource with a simple have_xft = 0.

Thanks to Garett Wilson, Raffaella Trainello
and Herman Robak for useful help.