Step 2: Updated all the XFT subsystem functions, on the way to managing UTF8
To prepare for utf8, I simply replaced the XFT*8 function with Xft*Utf8
and made a small change with the conversions of variables.

Note: this patch only works if the system is "X_HAVE_UTF8_STRING" enabled;
otherwise, go back to the previous code.

Thanks to Garett Wilson, Raffaella Trainello
and Herman Robak for useful help.