Step 4: Fixed the Utf8 deleting and cursor positioning problems
I solved the deleting and positioning problems with a little function.

Utf8 solves the problem of non ascii/iso 7bit (0x00 - 0x7F) with a
sequence of 1-6 8bit with the first byte higher than 0x80. I solved this
with a little function that, if it detects utf8, exits with the exact
positions to be skipped/deleted (also with the reverse direction).

Thanks to Garett Wilson, Raffaella Trainello
and Herman Robak for useful help.