Added ffmpeg 0.6-0.8 patch
This patch updated ffmpeg patch that supports both the included ffmpeg
version and external ffmpeg.
This patch also fixes issues with fileac3.C.

This patch appears to work on both 2.1.5 and 2.2 versions
of Cinelerra.

Added some fix on configure.in and ffmpeg.C to build with internal
ffmpeg 0.8.
Paolo Rampino