Complete new overlay engine and resampler for CinelerraCV
This is a new overlay engine resampler intended as a drop-in replacement for the older,
somewhat blurrier engine.  It also implements one new resampling mode (Lanczos) and adds
this to the preferences menu.

The titler plugin subclassed internals of the old resampler; the necessary lookup table generation
hs been moved into the plugin.

Improvements:

- Implements resampling in one step rather than scale + linear translation
- Considerably sharper in all modes
- Nearest Neighbor mode no longer shifts pixels to one side
- Added Lanczos for users that need even sharper output
- Border handling improved; eg, HDV can now be expanded exactly to full frame without semitransparent edges.
- Alpha blending has been changed to a 'progressive opaqueness' model rather than simply taking the alpha from the most opaque layer.

There's still plenty of room for improvement/optimization, but this appears to be ready for basic stress testing.