Initialize struct tm in dcraw.c before calling mktime()
When running cinelerra in valgrind (after building with -O0 to avoid
false positives from optimized string functions) and loading an MJPEG
AVI file, valgrind report that mktime() in dcraw.c do its calculations
using uninitialized memory.  This patch the code to make sure all fields
in the tm struct is initialized to a known value before mktime() try to
use it.  Tested and found to not affect the file loading, while getting
rid of the valgrind error.