Make sure rgba values always have known values in overlayframe.C
Coverity observed that it could happen that the BLEND_3() and BLEND_4()
macros used by the TRANSLATE() macro could end up setting the rgb(a)
values to uninitialized content if the mode is not among the values
handled by the switch().  This patch make sure also this case result in
well known values - setting them to zero to get black/transparent as the
result if this ever happen, and reduce the amount of open Coverity
issues.  An alternative which also would make sense, would be to set
them to the max value to get white/non-trasparent.  I picked black
fairly randomly, as zero seemed like a nice value to use for
initialization.

Fixes Coverity CIDs 58909-58912.