Move AffineEngine implementation from plugins/motion/ to cinelerra/.
When the findobject plugin was merged from HV 4.5, the implementation
of the AffineEngine class was included in the cinelerray binary.  This
implementation proved to conflict with the original implementation in
plugins/motion/libaffine.a, causing plugins to misbehave and cinelerra
to crash (John Griffiths reported a crash in the rotate plugin).

Remove the old and now redundant implemetation in the motion plugin to
bring us closer to HV 4.5, after checking that no CV related patch was
missing in the new version in cinelerra/affine.C.  Use the new copy
in cinelerra instead.

Added two compatibility functions AffineEngine::set_viewport() and
AffineEngine::set_pivot() to avoid having to rewrite plugins to use
the new functions where the operation was split into input and output
viewport and pivot.  These compatibility functions should be removed
when the plugins are updated from HV 4.5.