Backport fileffmpeg from cinelerra4.5
This allows to open unknow files directly with ffmpeg
All it's handle by fileffmpeg, fixed also conflicts with
filevorbis.
On my tests I noticed that this metod has serious problem with color
space and probably memory leaks. Needs to be fixed.

For this reasons, for now, it handles files instead default pcm.