Gtk3 file chooser for cinelerra
This patch adds possibility to substitute the default file loader of cinelerra with
gtk3 file chooser. To activate you have to configure with --enable-gtk.
To accomplish this I had to write a wrapper ( gtkwrapper ) to isolate glib
from guicast. I'm sorry for complexity of this patch, but it not was possible
to do in other way.

Warning very experimental.

Known (and strange) issue: if on gtk chooser you press cancel and then exit from
cinelerra, the preference screen format will be overwritten with compressed width
settings.